Remove mongoose, mongo and jasmine files
