+++ app.js
@@ -1,0 +1,12 @@
+const express = require('express')
+const app = express()
+
+app.use(express.static(__dirname + '/public'))
+
+app.get('/', function (req, res) {
+  res.sendFile(__dirname + '/public/home.html')
+})
+
+app.listen(3000, function () {
+  console.log('Example app listening on port 3000!')
+})
