--- route/index.js
@@ -1,1 +1,0 @@
-
